app/discord-to-telegram/requirements.txt
discord.py-self==1.9.2
python-telegram-bot==13.14
